Give Network and Tweaks independent Python CLIs
