+++ .github/workflows/mac.yml
@@ -11,7 +11,8 @@ on:
       - 'tools/tweaks/frontend/**'
       - 'tools/network/frontend/**'
       - 'contracts/**'
-      - 'cli/snapo'
+      - 'skills/snap-o-network-inspector/scripts/**'
+      - 'skills/snap-o-tweaks/scripts/**'
       - 'tool-reader/**'
       - '.github/workflows/mac.yml'
       - 'VERSION'
@@ -22,7 +23,8 @@ on:
       - 'tools/tweaks/frontend/**'
       - 'tools/network/frontend/**'
       - 'contracts/**'
-      - 'cli/snapo'
+      - 'skills/snap-o-network-inspector/scripts/**'
+      - 'skills/snap-o-tweaks/scripts/**'
       - 'tool-reader/**'
       - '.github/workflows/mac.yml'
       - 'VERSION'
