+++ AGENTS.md
@@ -3,10 +3,10 @@ Snap-O is a public, open-source Android inspection tool.
 ## Repository Layout
 
 - `app-macos/`: macOS app, device transport, and native tests.
-- `cli/`: standalone Python client and its tests.
-- `tools/`: Network and Tweaks, each with Android libraries and a frontend.
+- `tools/`: Network and Tweaks, each with Android libraries, a frontend, and CLI tests.
 - `tool-sdk/`: tool plugin authoring APIs, runtime, and Gradle integration.
-- `tool-reader/`: Android helper used by the macOS app and CLI to read tool plugin metadata and frontend assets.
+- `tool-reader/`: Android helper used by the macOS app to read tool plugin metadata and frontend assets.
+- `skills/`: installable Network and Tweaks skills, each bundling its standalone Python CLI.
 - `examples/`: Android demo apps, an independent tool plugin project, and CLI examples.
 - `build-logic/` and `gradle/`: internal Gradle conventions and root build tooling.
 - `contracts/`: shared protocol definitions and fixtures.
@@ -18,8 +18,8 @@ Snap-O is a public, open-source Android inspection tool.
 - Read the component's `AGENTS.md` before changing it.
 - Keep macOS ADB and device transport in `app-macos/Snap-O/Device/`. Keep UI code in the app and tool plugin frontends.
 - Android tool plugins use `tool-sdk/runtime` and its `ToolServer` API. Keep shared socket and HTTP behavior in the runtime.
-- The Python CLI implements its own transport; it does not use the Swift client.
-- When changing shared behavior, check the Android libraries, macOS client, web tool, and Python CLI. Use `contracts/` and its fixtures to keep them compatible.
+- Each Python CLI implements its own transport; it does not use the Swift client.
+- When changing shared behavior, check the Android libraries, macOS client, web tool, and both Python CLIs. Use `contracts/` and its fixtures to keep them compatible.
 
 ## Public Repository
