+++ release/preflight.sh
@@ -309,11 +309,11 @@ print_client_protocol_declarations() {
       'snapo-network-inspector-web/src/features/tweaks-inspector/TweaksInspectorApp.tsx' \
       'inspectors/tweaks/src/features/tweaks-inspector/TweaksInspectorApp.tsx'
   fi
-  if git -C "$SNAPO_DIR" grep -q -E '^TWEAKS_PROTOCOL_VERSION[[:space:]]*=' "$1" -- cli/snapo scripts/snapo; then
+  if git -C "$SNAPO_DIR" grep -q -E '^TWEAKS_PROTOCOL_VERSION[[:space:]]*=' "$1" -- skills/snap-o-tweaks/scripts/snapo-tweaks cli/snapo scripts/snapo; then
     print_protocol_declaration "$1" 'CLI Network supported version' \
-      '^NETWORK_PROTOCOL_VERSION[[:space:]]*=' 'cli/snapo' 'scripts/snapo'
+      '^NETWORK_PROTOCOL_VERSION[[:space:]]*=' 'skills/snap-o-network-inspector/scripts/snapo-network' 'cli/snapo' 'scripts/snapo'
     print_protocol_declaration "$1" 'CLI Tweaks supported version' \
-      '^TWEAKS_PROTOCOL_VERSION[[:space:]]*=' 'cli/snapo' 'scripts/snapo'
+      '^TWEAKS_PROTOCOL_VERSION[[:space:]]*=' 'skills/snap-o-tweaks/scripts/snapo-tweaks' 'cli/snapo' 'scripts/snapo'
   else
     print_protocol_declaration "$1" 'CLI Tweaks minimum-version checks' \
       'protocol_version[[:space:]]*<[[:space:]]*[0-9]+' \
@@ -359,7 +359,7 @@ print_protocol_evidence 'Android servers' "$ANDROID_BASE" \
   contracts tool-sdk tool-reader tools sdk plugin-reader build-logic gradle build.gradle.kts settings.gradle.kts gradle.properties examples/android plugins/discovery android-discovery scripts/snapo-discovery.jar snapo-link-android plugins
 print_protocol_evidence 'Mac/web/CLI clients' "$MAC_BASE" \
   print_client_protocol_declarations \
-  contracts app-macos cli tool-sdk/host tool-reader examples/tool tools/network/frontend tools/tweaks/frontend sdk/host plugin-reader examples/plugin plugins/tweaks/frontend plugins/discovery android-discovery snapo-app-mac/SnapODeviceClient snapo-app-mac/Snap-O/NetworkInspector \
+  contracts app-macos cli skills tool-sdk/host tool-reader examples/tool tools/network/frontend tools/tweaks/frontend sdk/host plugin-reader examples/plugin plugins/tweaks/frontend plugins/discovery android-discovery snapo-app-mac/SnapODeviceClient snapo-app-mac/Snap-O/NetworkInspector \
   snapo-app-mac/Snap-O/Inspectors snapo-app-mac/Snap-O/Tools snapo-network-inspector-web inspectors tools plugin-host plugins/host snapo-link-android/tweaks-core/frontend snapo-link-android/network/frontend plugins/tweaks-core/frontend plugins/network/frontend plugins/implementations/tweaks/frontend plugins/implementations/network/frontend scripts snapo-link-android/example plugins/example
 printf '%s\n' 'Protocol review must be recorded for this source SHA before the version bump; this report does not approve compatibility.'
